\name{bsp.graph.fmem}
\alias{bsp.graph.fmem}
\title{Tool for plotting the nonlinear effects that are approximated by using B-splines for flexible measurement error models}
\description{
\bold{bsp.graph.fmem} displays the graphs of the nonparametric effects from an object of
the class \code{fmem()}.
}
\usage{
bsp.graph.fmem(object, which, xlab, ylab, main)
}
\arguments{
  \item{object}{An object of the class \code{fmem()}.}
  \item{which}{An integer value, that indicates which nonparametric effect is required.}
  \item{xlab}{(optional) A title for the \emph{x} axis}
  \item{ylab}{(optional) A title for the \emph{y} axis}
  \item{main}{(optional) An overall title for the graph.}
}
\author{Luz Marina Rondon <lumarp@gmail.com> and Heleno Bolfarine}


