\name{bsp.graph}
\alias{bsp.graph}
\title{Tool for plotting B-splines}
\description{
\bold{bsp.graph} displays the graphs of the nonparametric effects from an object of
the class \code{gesm()}.
}
\usage{
bsp.graph(object, which, xlab, ylab, main)
}
\arguments{
  \item{object}{An object of the class \code{gesm()} or \code{fmem()}.}
  \item{which}{An integer value, where 1 indicates location submodel, and 2 indicates dispersion submodel.}
  \item{xlab}{(optional) A title for the \emph{x} axis}
  \item{ylab}{(optional) A title for the \emph{y} axis}
  \item{main}{(optional) An overall title for the graph.}
}
\author{Luz Marina Rondon <lumarp@gmail.com> and Heleno Bolfarine}


