\name{summary.BayesGESM}
\alias{summary.BayesGESM}
\title{summary.BayesGESM}
\description{
\bold{summary.BayesGESM} displays the summary of the Bayesian analysis of Generalized Elliptical Semiparametric Models.
This function produces a table with the summary statistics (mean, median and standard deviation) of the posterior distribution
as well as the 95\% credible interval for the interest parameters. Further, this function displays goodness-of-fit statistics
such as DIC, EAIC, EBIC and LMPL.}
