\name{summary.fmem}
\alias{summary.fmem}
\title{Produces a complete summary of flexible measurement error model fit}
\description{
\bold{summary.fmem} displays the summary of the Bayesian analysis of flexible measurement error models.
This function produces a table with the summary statistics (mean, median and standard deviation) of the posterior distribution
as well as the 95\% credible interval for the interest parameters. Further, this function displays goodness-of-fit statistics
such as DIC and LMPL.}