\name{summary.gesm}
\alias{summary.gesm}
\title{Produces a complete summary of Generalized elliptical semi-parametric model fit}
\description{
\bold{summary.gesm} displays the summary of the Bayesian analysis of Generalized Elliptical Semi-parametric Models.
This function produces a table with the summary statistics (mean, median and standard deviation) of the posterior distribution
as well as the 95\% credible interval for the interest parameters. Further, this function displays goodness-of-fit statistics
such as DIC and LMPL.}
