\name{bsp}
\alias{bsp}
\title{Tool to approximate smooth functions by B-splines.}
\description{
\bold{bsp} is used to approximate smooth functions by B-splines.}
\usage{
bsp(x)
}
\arguments{
  \item{x}{values of the explanatory variable.}
}
\value{
\item{x}{the B-spline basis matrix, which is cubic and has an intercept.}
}
\references{De Boor, C. (1978). A practical Guide to Splines. Applied Mathematical Sciences. Springer-Verlag, New York.}
\author{Luz Marina Rondon <lumarp@gmail.com> and Heleno Bolfarine}